+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support testkit firmware 3.60
 - Added "Restart" text to poweroff button (20 languages)
 - Configurable Quick Menu PS button push time
+- Option to restart with the standby button
 
 ### Changelog - Quick Power